+++ app/helpers/agilibox/i18n_helper.rb
@@ -1,0 +1,5 @@
+module Agilibox::I18nHelper
+  def ta(action)
+    t("actions.#{action}")
+  end
+end