+++ spec/helpers/agilibox/i18n_helper_spec.rb
@@ -1,0 +1,9 @@
+require 'rails_helper'
+
+describe Agilibox::I18nHelper, type: :helper do
+  describe "ta" do
+    it "should translate action" do
+      expect(ta :update).to eq "Modifier"
+    end
+  end
+end